RPM deep scanner and fixes
