+++ .github/workflows/action.yml
@@ -1,11 +1,9 @@
-# This workflow tests github action feature and integration
+# This workflow tests github action feature and integration only for push action
 
 name: ActionTest
 
 on:
   push:
-    branches: [ main ]
-  pull_request:
     branches: [ main ]
 
 jobs:
