+++ .mypy.ini
@@ -52,3 +52,6 @@ ignore_missing_imports = True
 
 [mypy-base62.*]
 ignore_missing_imports = True
+
+[mypy-rpmfile.*]
+ignore_missing_imports = True
