+++ docs/source/credsweeper.deep_scanner.rst
@@ -124,6 +124,14 @@ credsweeper.deep\_scanner.mxfile\_scanner module
    :undoc-members:
    :show-inheritance:
 
+credsweeper.deep\_scanner.patch\_scanner module
+-----------------------------------------------
+
+.. automodule:: credsweeper.deep_scanner.patch_scanner
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 credsweeper.deep\_scanner.pdf\_scanner module
 ---------------------------------------------
 
@@ -144,6 +152,14 @@ credsweeper.deep\_scanner.pptx\_scanner module
 ----------------------------------------------
 
 .. automodule:: credsweeper.deep_scanner.pptx_scanner
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+credsweeper.deep\_scanner.rpm\_scanner module
+---------------------------------------------
+
+.. automodule:: credsweeper.deep_scanner.rpm_scanner
    :members:
    :undoc-members:
    :show-inheritance:
