+++ pyproject.toml
@@ -25,6 +25,7 @@ dependencies = [
     "python-docx",
     "python-pptx",
     "PyYAML",
+    "rpmfile",
     "whatthepatch",
     "xlrd",
 ]
