+++ requirements.txt
@@ -39,6 +39,7 @@ python-dateutil==2.9.0.post0
 python-docx==1.1.2
 python-pptx==1.0.2
 PyYAML==6.0.2
+rpmfile==2.1.0
 whatthepatch==1.0.7
 
 # Auxiliary
